Refactor PSI hooks and centralize API calls
